+++ src/lib/fetchApi.js
@@ -1,6 +1,7 @@
 import 'whatwg-fetch';
 
-const fetchApi = endpoint =>
-  fetch(`/api/${endpoint}`).then(response => response.json());
+const fetchApi = endpoint => fetch(`/api/${endpoint}`).then(response => response.json());
+
+export const fetchUrl = url => fetch(url).then(response => response.json());
 
 export default fetchApi;