+++ src/reducers/generic/generic.js
@@ -1,0 +1,16 @@
+import { handleActions } from 'redux-actions';
+import onReceive from '../../actions/generic/generic';
+
+const reducer = handleActions(
+  {
+    [onReceive]: (state, { payload }) => ({
+      ...state,
+      [payload.key]: payload,
+    }),
+  },
+  {
+    results: {},
+  },
+);
+
+export default reducer;